split rdb class into updater and queryer classes
